backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.5.0
tzdata==2024.1
